Refactor unionAuthors out of colors.ts
