+++ prototype1/src/const.ts
@@ -1,2 +1,8 @@
 export const padding = 30;
 export const textSpacingFromCircle = 5;
+export const users = [
+    ["joglr", "Jonas Glerup Røssum", "Jonas Røssum"],
+    ["tjomson", "Thomas Hoffmann Kilbak", "Thomas Kilbak"],
+    ["hojelse", "Kristoffer Højelse"],
+    ["emiljapelt", "Emil Jäpelt"],
+]
